% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrv_large.R
\name{abbrv_large}
\alias{abbrv_large}
\title{Abbreviated labels for \href{https://en.wikipedia.org/wiki/Names_of_large_numbers}{large numbers}.}
\usage{
abbrv_large(x)
}
\description{
`abbrv_large` takes a numeric vector and returns a character vector containing
 abbreviated names of the form `DDDC` where `DDD` are the three most significant
 digits of the number, and `C` an abbreviation which summarises its magnitude.
 For instance, `1e6` will be mapped to `1M`, and `123845` to `123k`. Note
 that numbers are _truncated_, and not rounded, to obtain `DDD`.
}
\examples{

 abbrv_large(c(123456, 28475777, 125, 1.3, 0.1, 1))

}
