% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directlabels.R
\name{bumpup_bounded}
\alias{bumpup_bounded}
\title{Bump labels up, but drops some if too crammed}
\usage{
bumpup_bounded(d, max_bump = 1.1, ...)
}
\arguments{
\item{d}{the directlabels positioning dataframe.}

\item{max_bump}{the maximum bump factor allowed before a label is dropped.
Expressed as a factor of the label's height.}
}
\description{
Label positioning strategy similar to \code{\link{bumpup}}, but with a cap
on the maximum vertical displacement that can be applied to a label. If
the required displacement overshoots the cap, the label is dropped. This
creates an effect similar to the labelling strategy seen in Tableau,
except with better results.
}
