% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_playax.R
\name{palette_playax}
\alias{palette_playax}
\alias{palette_playax_unassigned}
\alias{palette_playax_all}
\alias{palette_playax_select}
\alias{palette_playax_extended}
\alias{palette_playax_services}
\title{A palette with a set of predefined colors}
\usage{
palette_playax()

palette_playax_unassigned()

palette_playax_all(assignment = c())

palette_playax_select(services = c())

palette_playax_extended(extension = c(), replacements = c())

palette_playax_services()
}
\arguments{
\item{extension}{a character vector with extra service names.}
}
\description{
`palette_playax_XX` exposes a set of predefined colors which are either _assigned_
to existing streaming services (which can be checked with `palette_playax_services`)
or are otherwise _unassigned_ and can be reused for other purposes.
}
\details{
These colors should be used in time series, histograms, and any plots where their corresponding
streaming service arises, and completed with the unassigned colors if there
is a need to present extra information (e.g. data from other services).

See \link{scale_color_playax} on how to access the Playax palette from
`ggplot2`.

\describe{
   \item{`palette_playax()`}{Returns a named character vector where
   each name corresponds to a streaming service (e.g. `spotify`) and each value
   corresponds to an RGB color string.}
   \item{`palette_playax_unassigned()`}{Returns an unnamed character vector
   containing RGB color strings which are not assigned to any service.}
   \item{`palete_playax_all()`}{Returns a character vector with all
   of the colors in the palette, assigned or otherwise. Clients may
   optionally specify custom assignment names for the colors.}
   \item{`palette_playax_extended()`}{Returns a named character vector with
   all of the colors in `palette_playax_assigned()`, plus a set of unassigned
   colors picked from the palette and bound to the names passed in the
   `extension` parameter.}
   \item{`palette_playax_services()`}{Returns a character vector with the
   set of streaming services supported by the palette.}
   \item{`palette_playax_select()`}{Returns colors assigned to specific
   services. Shortcut for `palette_playax()[services].` }
}
}
