% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playaxr.R
\docType{package}
\name{playaxr-package}
\alias{playaxr}
\alias{playaxr-package}
\title{playaxr: tools for building standards-compliant plots}
\description{
playaxr provides a set of tools for building plots which comply
to Playax's standards. These include predefined palettes for displaying
music streaming services, as well as formatting tools for displaying play
counts and more.
}
