% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{scale_fill_playax}
\alias{scale_fill_playax}
\alias{scale_color_playax}
\alias{scale_y_abbrv}
\alias{scale_x_abbrv}
\title{scales for Playax plots in \link{ggplot2}}
\usage{
scale_fill_playax(..., aesthetics = "fill", palette = palette_playax())

scale_color_playax(..., aesthetics = "color",
  palette = palette_playax())

scale_y_abbrv(...)

scale_x_abbrv(...)
}
\description{
This package provides:
\itemize{
  \item `color` and `fill`` scales based on \link{palette_playax};
  \item `x` and `y` scales which abbreviate large numbers with \link{abbr_large}.
}
}
\examples{


library(ggplot2)

data <- data.frame(
  day = rep(1:20, 3),
  metric = do.call(c, lapply(1:3, function(x) cumsum(rnorm(20, sd = 1e7)))),
  service = rep(c('facebook', 'instagram', 'spotify'), each = 20)
)

ggplot(data) +
  geom_line(aes(x = day, y = metric, col = service)) +
  theme_playax() +
  scale_color_playax() +
  scale_y_abbrv()

data <- rbind(
  data,
  data.frame(
    day = 1:20,
    metric = cumsum(rnorm(20, sd = 1e7)),
    service = rep('other', 20)
  )
)

ggplot(data) +
  geom_line(aes(x = day, y = metric, col = service)) +
  theme_playax() +
  scale_color_playax(palette = palette_playax_extended(extension = c('other'))) +
  scale_y_abbrv()

}
