% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_playax.R
\name{theme_playax}
\alias{theme_playax}
\title{Playax theme for ggplot2 charts}
\usage{
theme_playax()
}
\description{
This theme requires Google's Open Sans fonts
\url{https://fonts.google.com/specimen/Open+Sans} to work properly.
}
\examples{

library(ggplot2)

ggplot(tibble(x = 1:30, y = rnorm(30))) +
  geom_line(aes(x = x, y = y)) +
  theme_playax()

}
