% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_playax.R
\name{theme_playax_black}
\alias{theme_playax_black}
\title{Playax "black" theme for ggplot2 charts}
\usage{
theme_playax_black()
}
\description{
This theme requires Google's Open Sans fonts
\url{https://fonts.google.com/specimen/Open+Sans} to work properly.
}
